Further refinement to conversion -- proper eta for etop
Use eta to determine etop. This is what PTM does, and even
though etop is an interval value, and eta is instantaneous,
it's not correct to use eta_avg to get etop.  The precise
etop is eta_max over the interval. max(eta,eta_avg) works
okay, but for closest consistency with PTM, just use eta.